Moves hash_to_curve into crypto_hash
