+++ algorithms/src/commitment/pedersen.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{crh::PedersenCRH, hash_to_curve::hash_to_curve, CommitmentError, CommitmentScheme, CRH};
+use crate::{crh::PedersenCRH, crypto_hash::hash_to_curve, CommitmentError, CommitmentScheme, CRH};
 use snarkvm_curves::{AffineCurve, ProjectiveCurve};
 use snarkvm_fields::{ConstraintFieldError, Field, PrimeField, ToConstraintField};
 use snarkvm_utilities::BitIteratorLE;
