+++ algorithms/src/crypto_hash/mod.rs
@@ -17,6 +17,9 @@
 pub mod blake2xs;
 pub use blake2xs::*;
 
+pub mod hash_to_curve;
+pub use hash_to_curve::*;
+
 pub mod poseidon;
 pub use poseidon::*;
