+++ algorithms/src/lib.rs
@@ -44,9 +44,6 @@ pub use errors::*;
 #[cfg(feature = "fft")]
 pub mod fft;
 
-#[cfg(feature = "hash_to_curve")]
-pub mod hash_to_curve;
-
 #[cfg(feature = "merkle_tree")]
 pub mod merkle_tree;
